azure-openai-webapp/requirements.txt
flask
python-dotenv
openai
azure-identity
python-decouple
gunicorn
